Extract views to own files. Add previews
