+++ WooCommercePOS/WooCommercePOS/Models/CartProduct.swift
@@ -1,0 +1,5 @@
+
+struct CartProduct {
+    let product: Product
+    let quantity: Int
+}
