+++ WooCommercePOS/WooCommercePOS/Presentation/ProductCardView.swift
@@ -1,0 +1,37 @@
+import SwiftUI
+
+struct ProductCardView: View {
+    private let product: Product
+    private var onProductCardTapped: (() -> Void)?
+
+    init(product: Product, onProductCardTapped: (() -> Void)? = nil) {
+        self.product = product
+        self.onProductCardTapped = onProductCardTapped
+    }
+
+    var body: some View {
+        Rectangle()
+            .fill(.blue)
+            .frame(minWidth: 200, maxWidth: .infinity, minHeight: 200)
+            .aspectRatio(1, contentMode: .fit)
+            .overlay {
+                VStack {
+                    Text(product.name)
+                        .foregroundStyle(Color.cyan)
+                    Text(product.price)
+                        .foregroundStyle(Color.cyan)
+                    HStack {
+                        QuantityBadgeView()
+                        PlusButtonView {
+                            onProductCardTapped?()
+                        }
+                    }
+                }
+            }
+    }
+}
+
+#Preview {
+    ProductCardView(product: Product(itemID: UUID(), productID: 0, name: "Product name", price: "$2.00"),
+                    onProductCardTapped: { })
+}
