+++ WooCommercePOS/WooCommercePOS/Presentation/ProductGridView.swift
@@ -1,64 +1,5 @@
 import SwiftUI
 
-struct PlusButtonView: View {
-
-    private var buttonTapped: (() -> Void)?
-
-    init(buttonTapped: (() -> Void)? = nil) {
-        self.buttonTapped = buttonTapped
-    }
-
-    var body: some View {
-        Button(action: {
-            buttonTapped?()
-        }) {
-            Image(systemName: "plus")
-                .font(.system(size: 24, weight: .bold))
-                .foregroundColor(.white)
-                .frame(width: 50, height: 50)
-                .background(Circle().fill(Color.clear))
-        }
-    }
-}
-
-struct QuantityBadgeView: View {
-    var body: some View {
-        Text("2")
-    }
-}
-
-struct ProductCardView: View {
-
-    private let product: Product
-    private var onProductCardTapped: (() -> Void)?
-
-    init(product: Product, onProductCardTapped: (() -> Void)? = nil) {
-        self.product = product
-        self.onProductCardTapped = onProductCardTapped
-    }
-
-    var body: some View {
-        Rectangle()
-            .fill(.blue)
-            .frame(minWidth: 200, maxWidth: .infinity, minHeight: 200)
-            .aspectRatio(1, contentMode: .fit)
-            .overlay {
-                VStack {
-                    Text(product.name)
-                        .foregroundStyle(Color.cyan)
-                    Text(product.price)
-                        .foregroundStyle(Color.cyan)
-                    HStack {
-                        QuantityBadgeView()
-                        PlusButtonView {
-                            onProductCardTapped?()
-                        }
-                    }
-                }
-            }
-    }
-}
-
 struct ProductGridView: View {
     @ObservedObject var viewModel: PointOfSaleDashboardViewModel
