+++ WooCommercePOS/WooCommercePOS/Presentation/QuantityBadgeView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct QuantityBadgeView: View {
+    var body: some View {
+        Text("2")
+    }
+}
+
+#Preview {
+    QuantityBadgeView()
+}
